# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Tracy API
  version: 0.1.0
  description: Self-hosted LLM and agent trace observability API.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/me:
    get:
      operationId: currentUser
      responses:
        '200':
          description: Current user and active Workspace.
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/oauth/apps:
    get:
      operationId: listOAuthApps
      description: Admin-only OAuth App list.
      responses:
        '200': { description: OAuth Apps. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createOAuthApp
      description: Registers an RSA public key for a CozeLoop JWT OAuth client.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [client_id, project_id, public_key_id, public_key]
              properties:
                client_id: { type: string }
                project_id: { type: string }
                public_key_id: { type: string }
                public_key: { type: string, description: RSA PEM public key }
                enabled: { type: boolean, default: true }
      responses:
        '201': { description: OAuth App registered. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    BadRequest:
      description: Invalid request or query.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Admin API key required.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
...